# improve perf of asString