--- a/lib/serializer.js
+++ b/lib/serializer.js
@@ -113,11 +113,10 @@ module.exports = class Serializer {
       }
     }
 
-    // Fast escape chars check
-    if (!STR_ESCAPE.test(str)) {
-      return '"' + str + '"'
-    } else if (str.length < 42) {
+    if (str.length < 42) {
       return this.asStringSmall(str)
+    } else if (STR_ESCAPE.test(str) === false) {
+      return '"' + str + '"'
     } else {
       return JSON.stringify(str)
     }
@@ -130,32 +129,32 @@ module.exports = class Serializer {
   // 34 and 92 happens all the time, so we
   // have a fast case for them
   asStringSmall (str) {
-    const l = str.length
+    const len = str.length
     let result = ''
-    let last = 0
-    let found = false
-    let surrogateFound = false
+    let last = -1
     let point = 255
+
     // eslint-disable-next-line
-    for (var i = 0; i < l && point >= 32; i++) {
+    for (var i = 0; i < len; i++) {
       point = str.charCodeAt(i)
+      if (point < 32) {
+        return JSON.stringify(str)
+      }
       if (point >= 0xD800 && point <= 0xDFFF) {
         // The current character is a surrogate.
-        surrogateFound = true
+        return JSON.stringify(str)
       }
-      if (point === 34 || point === 92) {
+      if (
+        point === 0x22 || // '"'
+        point === 0x5c // '\'
+      ) {
+        last === -1 && (last = 0)
         result += str.slice(last, i) + '\\'
         last = i
-        found = true
       }
     }
 
-    if (!found) {
-      result = str
-    } else {
-      result += str.slice(last)
-    }
-    return ((point < 32) || (surrogateFound === true)) ? JSON.stringify(str) : '"' + result + '"'
+    return (last === -1 && ('"' + str + '"')) || ('"' + result + str.slice(last) + '"')
   }
 
   getState () {