@mcollina 2023-07-26
> Why 42?

It's a reference to Douglas Adams' “The Hitchhiker's Guide to the Galaxy,” because that number is the answer given by a supercomputer to “the Ultimate Question of Life, the Universe, and Everything.”